feat: change architecture
