+++ services/video_streaming/app/config/go.mod
@@ -1,5 +1,3 @@
 module config
 
 go 1.20
-
-require github.com/rs/cors v1.9.0
