+++ services/video_streaming/app/config/initializers/config.go
@@ -1,4 +1,4 @@
-package config
+package initializers
 
 import (
 	"net/http"
