+++ services/video_streaming/app/config/initializers/go.mod
@@ -1,0 +1,5 @@
+module initializers
+
+go 1.20
+
+require github.com/rs/cors v1.9.0
