+++ services/video_streaming/app/config/initializers/server.go
@@ -1,0 +1,16 @@
+package initializers
+
+import(
+	"net/http"
+)
+
+func NewServer(config *Config) *http.Server{
+	srv := &http.Server{
+		Addr:         config.Addr,
+		Handler:      config.Handler,
+		ReadTimeout:  config.ReadTimeout,
+		WriteTimeout: config.WriteTimeout,
+		IdleTimeout:  config.IdleTimeout,
+	}
+	return srv
+}
