+++ services/video_streaming/app/library/logger_request.go
@@ -1,0 +1,10 @@
+package library
+
+import (
+	"fmt"
+	"time"
+)
+
+func LoggerRequest(addr string, method string, url string, duration time.Duration) {
+	fmt.Println("\nRequest { Elapsed: ", duration, " Method: ", method, " Endpoint:", url, "}")
+}
