+++ services/video_streaming/app/router/routes.go
@@ -1,0 +1,18 @@
+package router
+
+import(
+	"net/http"
+	resolvers "resolvers"
+)
+
+func NewRouter(siteMux *http.ServeMux) *Router {
+	st := &resolvers.Storage{Bucket: "files", Key: "video_id"}
+
+	router := Router{
+		server: siteMux,
+		requests: routerRequests{
+			"/videos": st.StreamVideoResolver,
+		},
+	}
+	return &router
+}
