+++ services/video_streaming/server.go
@@ -4,9 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
-	. "config"
 	. "router"
-	. "app"
+	. "initializers"
 )
 
 func main() {
